Removed unnecessary define blocks
There was a big block of variable definitions in config.groovy
because the way we access variables from jelly: it.variable
didn't work properly in groovy and had to be fetched in the
beginning of the file. By changning this to using my.variable
instead, it was possible to databind where the variables
were needed instead.

Change-Id: Iffd4c32f6cf54922d21f79eec666036bb630f913