Changed how the plugin handles a newly submitted configuration
The configure method no longer removes the existing filter before the
new properties are evaluated. This brought back an earlier issue
where the config page breaks if the user submits a faulty config
and restarts Jenkins. To avoid this, the filter is disabled if
it fails to initialize after restart. It's then up to the user to
go to the config page and submit a new config. A check that login.conf
exists was added to the configuration method so the plugin can
inform the user about it properly.

Change-Id: I46e8cdaf9fe62a9dd5fe478e6058c5117c379800