Exception in automatic login for accessing userContent
If the user tries to access http://jenkinsserver/userContent
anonymously, the plugin will try to login the user automatically,
even though the documentation for userContent says it shouldn't
be subject to any access controls.

This change makes an exception in the automatic login if the
user tries to access userContent.

Change-Id: Ia344a21ee0e41bab9576d3bce73d7152f0a32d13