[JENKINS-75504] Use MIT license in pom and LICENSE.md file
https://issues.jenkins.io/browse/JENKINS-75504 reports that a license
is not declared by the plugin in either its pom file or a separate
LICENSE file.

Since all the Java files include an MIT license and the Jenkins project
recommends the MIT license, it seems safe to add the MIT license to
the pom.xml file and to create a LICENSE.md file with the MIT license.
Both the pom file segment and the LICENSE.md segment were copied from
the plugin archetype.  The year of the copyright is based on the first
year recorded in the git history of the plugin.

The two forms of copyright statements in the Java source files are:

 * Copyright (c) 2014 Sony Mobile Communications Inc. All rights reserved.
 * Copyright (c) Red Hat, Inc.